# workflow-manager/api/workflow/swagger/OpenAPI/apidocs.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Workflow manager API spec",
    "version": "1.0"
  },
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/workflow/v1/domain/{domain}/base-image/{name}/version/{version}": {
      "delete": {
        "summary": "delete of base image can be triggered only with specific versions",
        "operationId": "BaseImageApi_DeleteBaseImage",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/workflowBaseImageDelResp"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "domain",
            "description": "domain under consideration",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "name",
            "description": "name of the base image",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "version",
            "description": "version of the base image",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "BaseImageApi"
        ]
      }
    },
    "/api/workflow/v1/domain/{domain}/base-images": {
      "get": {
        "summary": "List of available base images",
        "operationId": "BaseImageApi_ListBaseImages",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/workflowBaseImagesListResp"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "domain",
            "description": "domain under consideration",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "BaseImageApi"
        ]
      }
    },
...
    "/api/workflow/v1/domain/{domain}/project/{project}/base-images": {
      "get": {
        "summary": "List of available base images in a project for module creation",
        "operationId": "BaseImageApi_ListProjectBaseImages",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/workflowBaseImagesListResp"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "domain",
            "description": "domain under consideration",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "project",
            "description": "project under consideration",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "BaseImageApi"
        ]
      }
    },
...
  },
  "definitions": {
...
    "googlerpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
...
    "workflowBaseImageDelResp": {
      "type": "object"
    },
    "workflowBaseImagesListEntry": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "title": "name of the image"
        },
        "versions": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/workflowBaseImagesVersionInfo"
          },
          "title": "list of supported versions for this image"
        }
      }
    },
    "workflowBaseImagesListResp": {
      "type": "object",
      "properties": {
        "items": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/workflowBaseImagesListEntry"
          },
          "title": "list of items available in the response"
        }
      }
    },
    "workflowBaseImagesVersionInfo": {
      "type": "object",
      "properties": {
        "version": {
          "type": "string",
          "title": "version identifier of the image"
        },
        "desc": {
          "type": "string",
          "title": "description for the image"
        },
        "imageId": {
          "type": "string",
          "title": "base image id to be used"
        },
        "extRef": {
          "type": "string",
          "title": "external reference for the image"
        }
      }
    },
...
  },
...
}